If x509_path (when calling GetCOndorClient) is not explicitly set, try to find an voms proxy using voms-init-proxy.  Warn user if voms proxy could not be found, but continue.  (sometimes a voms proxy is not needed.)
If x509_path is set, check for file existence and warn if it doesn't exist.

Note that ATM the revised code doesn't copy the proxy into the initial_dir
(currently /scratch/$USER), but tells condor the path of original file.
I think the condor_submit should be able to access and retrieve the original proxy
file as given by the path.

Possibly there are reasons to copy the proxy file into the initial_dir that I'm not aware of.